requirements.txt
Flask==3.0.0
yandex-cloud-ml-sdk
python-dotenv
numpy
Pillow